Update internal-kista.md (#36)
* Update internal-kista.md

Added description: "Less is a command that allows you to read the content of a file. We suggest you copy the information from this file to your favourite text editor to keep it available while performing the next steps. To close the file, press q." because (based on my own behaviour) users not familiar with linux tend to follow your instructions quite instrumentally. I did not immediately understand that I had just opened a file when I performed the less command.

* Changed kista example from less to cat to simplify viewing kista transfer information with additional rewording

Co-authored-by: Sandor Zeestraten <sandor.zeestraten@ntnu.no>